requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
waitress==2.1.2
Flask-RESTful==0.3.10
tensorflow==2.13.0
pandas==2.0.2
pika==1.2.0
rabbitmq-pika-flask==1.2.33
psycopg2-binary==2.9.6
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
scikit-learn==1.3.0
yahoo-fin==0.8.9.1
webargs==8.3.0
matplotlib==3.7.2